Add u-root integration's infrastructure to webboot
Louis and I copied the entire integration directory from u-root into
webboot. Then, since multiboot was currently failing, and we knew we
were not using it as of right now, we changed the "TestMultiBoot()" to
"testMultibootInternal()."

Update the README.md for better instructions to receive a kernel and
QEMU for the integration testing.

Signed-off-by: Urvisha Patel <urvishapatel@google.com>