make wifi work better by adding needed tools; make it possible to skip dhcp
The wifi setup, for more complex cases, needs the wpa_supplicant and other tools.
This PR adds them.

This PR also allows the user to skip dhcp; if both ipv4 and ipv6 are false,
it will not call dhclient at all. For example,
webboot -ipv4=false -ipv6=false
will skip the dhclient step entirely. This is handy if you
want to set up the network by hand.

Default to building with -build=bb. This fixes several things: first,
it ensures everything will build; second, it makes boot way
faster; third, it makes things much smaller which makes it so
the little x230 will run them. I had the very real problem
that webboot booted and got a compile error. We don't want that
to happen.

You can verify these changes work. When you build and boot with this,
you can, for example, run
wifi some-essid password-for-essid
and it will work fine and start the wpa_supplicant to keep
IP addresses refreshed. Tested on my x230 on my home network.

Then you can
webboot -essid="" -ipv4=false -ipv6=false Tinycore

and watch it fetch tinycore.

Note that the idea of Must seems weird, but consider a
few real-world uses:
regexp.MustCompile
template.Must

These break the model of always returning errors to the caller.
They are intended for the kind of initialization case we use here.

That said, I want to fix u-root so that, e.g.
u-root -files cat

will properly find /bin/cat and use it. Then we can remove this temporary
function. But that will require a version spin of u-root.

Further:

webboot had features that made it hard to use. It could
not work if the WIFI network had a password.

It needed to use the wifi command, not iwconfig, to properly
setup such networks.

It needed to set the output of wifi up to go to os.Stdout/Stderr
else you could never tell if anything was working.

It needed to run the wifi setup command in background, since it
might run a long time.

These changes make it much easier to use in complicated WIFI
networks.

Tested on my x230 and it's much better.

I've also changed buildimage.go so it no longer needs to run as root
all the time. If setting up some commands fails it will tell you
what you need to do.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>