remove the framebuffer console (and other things that enable it)
We were unable to boot the tinycore kernel. Turns out we enabled
the framebuffer console and the tinycore kernel could not handle that.

Turn off anything related to graphical startup. While we're at it
turn off other options that bloat up the kernel, make it slow to build,
and possibly cause trouble for kexec.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>