build webboot standalone, not as part of u-root
The u-root churn has caused us a lot of trouble lately.
The problem is that we were building webboot as part of a u-root
build, and the fact that then u-root depended on webboot
which then depended on u-root was a headache.

We now build webboot as a standalone program, so it can use the
vendored u-root in this directory. This will end the mess we've had.

While we're at it, take away the auto-fetch of u-root and NiChrome.
It's a headache when offline.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>