Verify checksum in exec
After an ISO has finished downloading and if the distro selected has a
checksum available, VerifyChecksum is called.

If the checksum for the ISO does not match the correct checksum for that
distro, the user will be asked whether or not they want to proceed
anyway. If not, the user will be taken back to the list of distro
options to pick a new one to download.

If the checksum is incorrect, the message will show the correct checksum
and the checksum of the downloaded ISO. VerifyChecksum is changed to
return the calculated  checksum.

There is a new test to test displayChecksumPrompt. TestChecksum is
updated to test whether VerifyChecksum returns the calculated checksum.

Signed-off-by: Kelly Sun <sunkelly888@gmail.com>